feat: 사용자 조회 response 변환 수정
