+++ src/main/java/com/lxp/Main.java
@@ -1,13 +1,13 @@
 package com.lxp;
 
-import com.lxp.config.JDBCConnection;
-import com.lxp.config.TransactionManager;
+import com.lxp.config.ApplicationContext;
+import com.lxp.config.DatabaseInitializer;
 import com.lxp.util.CLIRouter;
 
 public class Main {
     public static void main(String[] args) {
-        TransactionManager.DatabaseInitializer.initialize();
-        CLIRouter router = JDBCConnection.ApplicationContext.getRouter();
+        DatabaseInitializer.initialize();
+        CLIRouter router = ApplicationContext.getRouter();
         router.start();
     }
 }
