+++ src/main/java/com/lxp/config/JDBCConnection.java
@@ -50,87 +50,4 @@ public class JDBCConnection {
         System.out.println(
                 "대기중인 커넥션 요청 수" + poolMXBean.getThreadsAwaitingConnection()); // 풀에서 연결을 기다리는 스레드 수를 가져옴
     }
-
-    public static class ApplicationContext {
-
-        private ApplicationContext() {}
-
-        // enrollment component
-        private static class EnrollmentDaoHolder {
-            private static final EnrollmentDao INSTANCE = new EnrollmentDao();
-        }
-
-        private static class EnrollmentServiceHolder {
-            private static final EnrollmentService INSTANCE = new EnrollmentService(getEnrollmentDao());
-        }
-
-        private static class EnrollmentControllerHolder {
-            private static final com.lxp.api.controller.EnrollmentController INSTANCE =
-                    new com.lxp.api.controller.EnrollmentController(getEnrollmentService());
-        }
-
-        public static EnrollmentDao getEnrollmentDao() {
-            return EnrollmentDaoHolder.INSTANCE;
-        }
-
-        public static EnrollmentService getEnrollmentService() {
-            return EnrollmentServiceHolder.INSTANCE;
-        }
-
-        public static com.lxp.api.controller.EnrollmentController getEnrollmentController() {
-            return EnrollmentControllerHolder.INSTANCE;
-        }
-
-        // user component
-        private static class UserDaoHolder {
-            private static final UserDao INSTANCE = new UserDao();
-        }
-
-        private static class UserServiceHolder {
-            private static final UserService INSTANCE = new UserService(getUserDao());
-        }
-
-        private static class UserControllerHolder {
-            private static final UserController INSTANCE = new UserController(getUserService());
-        }
-
-        private static UserDao getUserDao() {
-            return UserDaoHolder.INSTANCE;
-        }
-
-        private static UserService getUserService() {
-            return UserServiceHolder.INSTANCE;
-        }
-
-        private static UserController getUserController() {
-            return UserControllerHolder.INSTANCE;
-        }
-
-        // handleController component
-        private static class HandleControllerHolder {
-            private static final EnrollmentHandler INSTANCE =
-                    new EnrollmentHandler(getEnrollmentController());
-        }
-
-        public static EnrollmentHandler getHandleController() {
-            return HandleControllerHolder.INSTANCE;
-        }
-
-        private static class UserHandlerHolder {
-            private static final UserHandler INSTANCE = new UserHandler(getUserController());
-        }
-
-        public static UserHandler getUserHandler() {
-            return UserHandlerHolder.INSTANCE;
-        }
-
-        // router component
-        private static class RouterHolder {
-            private static final CLIRouter INSTANCE = new CLIRouter(getHandleController(), getUserHandler());
-        }
-
-        public static CLIRouter getRouter() {
-            return RouterHolder.INSTANCE;
-        }
-    }
 }
