+++ src/main/java/com/lxp/infrastructure/mapper/UserMapper.java
@@ -1,5 +1,6 @@
 package com.lxp.infrastructure.mapper;
 
+import com.lxp.api.dto.UserResponse;
 import com.lxp.infrastructure.row.UserRow;
 import com.lxp.service.query.UserView;
 
@@ -15,5 +16,16 @@ public class UserMapper {
                 userRow.createdAt(),
                 userRow.updatedAt()
         );
+    }
+
+    public static UserResponse toUserResponse(UserView userView) {
+            return new UserResponse(
+                    userView.email(),
+                    userView.name(),
+                    userView.activeStatus(),
+                    userView.role(),
+                    userView.createdAt(),
+                    userView.updatedAt()
+            );
     }
 }
