+++ src/main/java/com/lxp/service/UserService.java
@@ -1,8 +1,10 @@
 package com.lxp.service;
 
+import com.lxp.api.dto.UserResponse;
 import com.lxp.config.TransactionManager;
 import com.lxp.domain.user.enums.UserRole;
 import com.lxp.infrastructure.dao.UserDao;
+import com.lxp.infrastructure.mapper.UserMapper;
 import com.lxp.service.query.UserView;
 
 public class UserService {
@@ -12,10 +14,11 @@ public class UserService {
         this.userDao = userDao;
     }
 
-    public UserView getUserById(Long id) throws Exception {
+    public UserResponse getUserById(Long id) throws Exception {
         try {
             TransactionManager.beginTransaction();
-            return userDao.findById(TransactionManager.getConnection(), id).orElseThrow(() -> new IllegalStateException("존재하지 않는 사용자 입니다."));
+            UserView userView = userDao.findById(TransactionManager.getConnection(), id).orElseThrow(() -> new IllegalStateException("존재하지 않는 사용자 입니다."));
+            return UserMapper.toUserResponse(userView);
         } finally {
             TransactionManager.close();
         }
